Fix call to "mex" helper under MATLAB 8.3 (R2014a) and GNU/Linux.
There is a new LINKOPTS variable which governs the list of symbols to be
exported, so we set it to an empty value to have all symbols exported.
Previously we were overloading the LDFLAGS which was containing the list of
exported symbols.

(cherry picked from commit 9aabb82c7662205699dd8a3ca400d878835d6ab2)