Fix bug in dynare_identification.m that led to missing output and crashes if the initial parameterization was invalid
If the initial parameters did not solve, the code for displaying results and saving was not reached. The also led to later crashes with saving as the -append option was used and no file existed. If no valid draw can be found, the function exits in any case, so no additional check is needed.
(cherry picked from commit 4c6aad3d2d8ae58bbaac5870e6f107645fac63c5)