posterior_sampler: Fix naming error that crashes parallel estimation
(cherry picked from commit ec3fb76aa7794141dc72bbd5f151dcc7a837b8bb)

Fixed conflict. In branch 4.5 the modified line was not under the condition testing if the model has auxiliary variables, and the
name of the structure was correct. So it was not the same bug. I suppose that Dynare was eventually crashing with models without
auxiliary variables (when sending the *_set_auxiliary_variables routine to the remotes), but I did not test.