Make rool for trend_component_model in var_expectation.
Not yet working, a bug in the preprocessor remains to be fixed. The
preprocessor does not create the correct number of reduced form parameters
for VAR_EXPECTATION when the auxiliary model is a trend component model,
because it ignores the fact that the model has to be rewritten in levels.