Use relative paths in .gitmodules for projects hosted on git.dynare.org
The main advantage is that, when someone does a recursive clone through SSH,
all the submodules will be cloned through SSH.

Relative paths are also useful when submodules are in private repositories: in
that case, the CI runner will have access to the private submodule.

It is important to enforce the group name in the relative URL, so that clones
in personal namespaces still work as expected.

See: https://docs.gitlab.com/ee/ci/git_submodules.html