doc: using `desc_addname` doesn't allow for line breaks.
The method `desc_addname` adds arguments to the name of the node, and hence does not allow for line breaks. In the case of Dynare commands we don't run into a problem with this but we run into this problem with longer MATLAB/Octave commands.

For Dynare commands, we don't run into the problem with commands running over the page width as we use placeholders like `OPTIONS` since the order of the parameters passed to a command is irrelevant

For MATLAB/Octave commands, the order of parameters passed to a function is important, hence we cannot use a placeholder such as `OPTIONS`

closes #1688