Manual: in a few places, replace “Octave” by “GNU Octave”
This is a partial revert of 8fa4c483f9116fb9f1f39d87aa584d4a8de10afb.

Actually, the official name is “GNU Octave”, and it is easier to find it under
that name in search engines. So use the full name at a few prominent places,
and use the shorter “Octave” everywhere else.