Source package: no longer embed a copy of MathJax
The copy of MathJax that we were embedding was not source code, because it
contained minified Javascript. In particular, this is a problem for the
official Debian package.

We now strip the copy. Users compiling the HTML manual from the source tarball
will therefore get MathJax from a CDN (this is the default behaviour of
Sphinx).

(cherry picked from commit d98cf64195ffec270515c9f922b14b58484121f3)