Windows package: fix compilation of MEX files for Octave
The runners no longer have an octave binary in the PATH, so the configure
script fails. We pass a dummy value for the location of the Octave
binary (which is actually not run, because we don’t execute the testsuite).

(cherry picked from commit 7a966a3088e07c572ad318a849fab47cad75fc6a)