Build system: workaround for compilation under MSYS2
There is (on 2021-03-29) a problem in the -static-libgfortran linker option that
affects the Dynare++ binaries.
See https://github.com/msys2/MINGW-packages/issues/7023
and https://github.com/msys2/MINGW-packages/issues/6986
Until this is fixed, the workaround is to add the --default-image-base-low linker
option.
We do *not* want to do this when cross-compiling from Debian (because the option does
not exist there), hence we test for cross-compilation before applying the workaround.