Fix spurious warning at startup under Octave about +pac/+bgp/{get,set}.m
This is a workaround for Octave bug 46849. It complements commit
4c0b2e8c4ec2df8f1833769db1c873486d2b87f7.

The workaround consists in moving the two problematic files to a subfolder
which is not in the path at startup. They will be added to the path later, at a
time when the Octave:shadowed-function warning has already been disabled by
warning_config.m.

Closes: #1817