Windows package: add -static-libgfortran when compiling MEX files for Octave
The Octave package for Windows ships with libgfortran from gfortran 9. This
version of libgfortran does not include symbol _gfortran_os_error_at, which is
needed by binaries created with gfortran 10 (used on our runners). The solution
is to statically link against libgfortran (as we already do for libgcc and
libstdc++).