Make get_equation_number_by_tag more robust
The implementation was relying on the fact that in M_.equations_tags, all
equations have a name tag and they appear in the order given by equation
numbers. There is no guarantee that this will always be the case, so use a more
robust approach.

By the way, improve the implementation of get_equation_name_by_number.