Testsuite: fix issue with Octave in a Wayland desktop session
If under a Wayland desktop session, some graphics would be displayed on the
screen, despite Octave being run in the virtual X11 frame buffer (xvfb).
Moreover, some unexplained random crashes would occur.

This commit ensures that Wayland is never used when running the testsuite.

By the way, add a command to explain why xvfb-run is needed and the
--no-window-system option cannot be run (see commit
96346b77e5e39ece565fc19f9b404f62c352488d).