Bytecode / conditional forecast: another fix test over consistency of variable flipping
The method Evaluate::getCurrentBlockEndogenous() would always return an empty
vector, because of a weirdness in FBEGINBLOCK class (after deserialization, the
latter has “variables” member unset and the corresponding data is instead store
in “Block_Contain_”).

On the master branch, the problem has been solved differently (the weirdness in
the FBEGINBLOCK class has been removed).