🐛 Fix MCP solver in the presence of multiple complementarity conditions
The solver relies on a reordering of equations (eq_index) so that reordered
equations appear at the index of the endogenous to which they are associated
through the complementarity condition.

However, this reordering could be computed incorrectly in the presence of
multiple complementarity conditions (technically, eq_index was not necessarily
a permutation).

(cherry picked from commit 1fea7f3e966b5c7803d4905ced29207440d13d57)